chore: move more instructions to their own modules
